Add discriminator from the slide
