+++ satflow/models/gan/discriminators.py
@@ -1,90 +1,13 @@
 import functools
 import torch
 from torch import nn as nn
-from torch.nn import init
+
+from satflow.models.gan.common import get_norm_layer, init_net
 
 
 class Identity(nn.Module):
     def forward(self, x):
         return x
-
-
-def get_norm_layer(norm_type="instance"):
-    """Return a normalization layer
-
-    Parameters:
-        norm_type (str) -- the name of the normalization layer: batch | instance | none
-
-    For BatchNorm, we use learnable affine parameters and track running statistics (mean/stddev).
-    For InstanceNorm, we do not use learnable affine parameters. We do not track running statistics.
-    """
-    if norm_type == "batch":
-        norm_layer = functools.partial(nn.BatchNorm2d, affine=True, track_running_stats=True)
-    elif norm_type == "instance":
-        norm_layer = functools.partial(nn.InstanceNorm2d, affine=False, track_running_stats=False)
-    elif norm_type == "none":
-
-        def norm_layer(x):
-            return Identity()
-
-    else:
-        raise NotImplementedError("normalization layer [%s] is not found" % norm_type)
-    return norm_layer
-
-
-def init_weights(net, init_type="normal", init_gain=0.02):
-    """Initialize network weights.
-
-    Parameters:
-        net (network)   -- network to be initialized
-        init_type (str) -- the name of an initialization method: normal | xavier | kaiming | orthogonal
-        init_gain (float)    -- scaling factor for normal, xavier and orthogonal.
-
-    We use 'normal' in the original pix2pix and CycleGAN paper. But xavier and kaiming might
-    work better for some applications. Feel free to try yourself.
-    """
-
-    def init_func(m):  # define the initialization function
-        classname = m.__class__.__name__
-        if hasattr(m, "weight") and (
-            classname.find("Conv") != -1 or classname.find("Linear") != -1
-        ):
-            if init_type == "normal":
-                init.normal_(m.weight.data, 0.0, init_gain)
-            elif init_type == "xavier":
-                init.xavier_normal_(m.weight.data, gain=init_gain)
-            elif init_type == "kaiming":
-                init.kaiming_normal_(m.weight.data, a=0, mode="fan_in")
-            elif init_type == "orthogonal":
-                init.orthogonal_(m.weight.data, gain=init_gain)
-            else:
-                raise NotImplementedError(
-                    "initialization method [%s] is not implemented" % init_type
-                )
-            if hasattr(m, "bias") and m.bias is not None:
-                init.constant_(m.bias.data, 0.0)
-        elif (
-            classname.find("BatchNorm2d") != -1
-        ):  # BatchNorm Layer's weight is not a matrix; only normal distribution applies.
-            init.normal_(m.weight.data, 1.0, init_gain)
-            init.constant_(m.bias.data, 0.0)
-
-    print("initialize network with %s" % init_type)
-    net.apply(init_func)  # apply the initialization function <init_func>
-
-
-def init_net(net, init_type="normal", init_gain=0.02):
-    """Initialize a network: 1. register CPU/GPU device (with multi-GPU support); 2. initialize the network weights
-    Parameters:
-        net (network)      -- the network to be initialized
-        init_type (str)    -- the name of an initialization method: normal | xavier | kaiming | orthogonal
-        gain (float)       -- scaling factor for normal, xavier and orthogonal.
-        gpu_ids (int list) -- which GPUs the network runs on: e.g., 0,1,2
-
-    Return an initialized network.
-    """
-    init_weights(net, init_type, init_gain=init_gain)
-    return net
 
 
 def define_D(input_nc, ndf, netD, n_layers_D=3, norm="batch", init_type="normal", init_gain=0.02):
@@ -125,6 +48,8 @@ def define_D(input_nc, ndf, netD, n_layers_D=3, norm="batch", init_type="normal"
         net = NLayerDiscriminator(input_nc, ndf, n_layers_D, norm_layer=norm_layer)
     elif netD == "pixel":  # classify if each pixel is real or fake
         net = PixelDiscriminator(input_nc, ndf, norm_layer=norm_layer)
+    elif netD == "enhanced":
+        net = CloudGANDiscriminator(input_channels=input_nc, num_filters=ndf, num_stages=3)
     else:
         raise NotImplementedError("Discriminator model name [%s] is not recognized" % netD)
     return init_net(net, init_type, init_gain)
@@ -197,58 +122,6 @@ class GANLoss(nn.Module):
             else:
                 loss = prediction.mean()
         return loss
-
-
-def cal_gradient_penalty(
-    netD, real_data, fake_data, device, type="mixed", constant=1.0, lambda_gp=10.0
-):
-    """Calculate the gradient penalty loss, used in WGAN-GP paper https://arxiv.org/abs/1704.00028
-
-    Arguments:
-        netD (network)              -- discriminator network
-        real_data (tensor array)    -- real images
-        fake_data (tensor array)    -- generated images from the generator
-        device (str)                -- GPU / CPU: from torch.device('cuda:{}'.format(self.gpu_ids[0])) if self.gpu_ids else torch.device('cpu')
-        type (str)                  -- if we mix real and fake data or not [real | fake | mixed].
-        constant (float)            -- the constant used in formula ( ||gradient||_2 - constant)^2
-        lambda_gp (float)           -- weight for this loss
-
-    Returns the gradient penalty loss
-    """
-    if lambda_gp > 0.0:
-        if (
-            type == "real"
-        ):  # either use real images, fake images, or a linear interpolation of two.
-            interpolatesv = real_data
-        elif type == "fake":
-            interpolatesv = fake_data
-        elif type == "mixed":
-            alpha = torch.rand(real_data.shape[0], 1, device=device)
-            alpha = (
-                alpha.expand(real_data.shape[0], real_data.nelement() // real_data.shape[0])
-                .contiguous()
-                .view(*real_data.shape)
-            )
-            interpolatesv = alpha * real_data + ((1 - alpha) * fake_data)
-        else:
-            raise NotImplementedError("{} not implemented".format(type))
-        interpolatesv.requires_grad_(True)
-        disc_interpolates = netD(interpolatesv)
-        gradients = torch.autograd.grad(
-            outputs=disc_interpolates,
-            inputs=interpolatesv,
-            grad_outputs=torch.ones(disc_interpolates.size()).to(device),
-            create_graph=True,
-            retain_graph=True,
-            only_inputs=True,
-        )
-        gradients = gradients[0].view(real_data.size(0), -1)  # flat the data
-        gradient_penalty = (
-            ((gradients + 1e-16).norm(2, dim=1) - constant) ** 2
-        ).mean() * lambda_gp  # added eps
-        return gradient_penalty, gradients
-    else:
-        return 0.0, None
 
 
 class NLayerDiscriminator(nn.Module):
@@ -353,3 +226,41 @@ class PixelDiscriminator(nn.Module):
     def forward(self, input):
         """Standard forward."""
         return self.net(input)
+
+
+class CloudGANBlock(nn.Module):
+    def __init__(self, input_channels):
+        super().__init__()
+        self.conv = torch.nn.Conv2d(input_channels, input_channels * 2, kernel_size=(3, 3))
+        self.relu = torch.nn.ReLU()
+        self.pool = torch.nn.MaxPool2d(kernel_size=(2, 2))
+
+    def forward(self, x):
+        x = self.conv(x)
+        x = self.relu(x)
+        x = self.pool(x)
+        return x
+
+
+class CloudGANDiscriminator(nn.Module):
+    """Defines a discriminator based off https://www.climatechange.ai/papers/icml2021/54/slides.pdf"""
+
+    def __init__(self, input_channels: int = 12, num_filters: int = 64, num_stages: int = 3):
+        super().__init__()
+        self.conv_1 = torch.nn.Conv2d(
+            input_channels, num_filters, kernel_size=1, stride=1, padding=0
+        )
+        self.stages = []
+        for stage in range(num_stages):
+            self.stages.append(CloudGANBlock(num_filters))
+            num_filters = num_filters * 2
+        self.stages = torch.nn.Sequential(*self.stages)
+        self.flatten = torch.nn.Flatten()
+        self.fc = torch.nn.LazyLinear(1)  # Real/Fake
+
+    def forward(self, x):
+        x = self.conv_1(x)
+        x = self.stages(x)
+        x = self.flatten(x)
+        x = self.fc(x)
+        return x
